Add support for zerocopy fields (#11)
- Add new `zerocopy` decorator to `ingot` fields, which mark fields that
can read / written directly as bytes
- Add `Ipv4Addr` and `Ipv6Addr` to `ingot_types`
- Use these new types in `ingot` packet headers

This is a little tricky, because we lose explicit tracking of size
during macro evaluation! Instead, size becomes a known number of bits,
plus some number of zerocopy fields. We can still compute size at
compile time with a `const` expression (using `size_of`). The zerocopy
fields are guaranteed to be byte-aligned, so if we are byte-aligned when
they start, we remain byte-aligned afterwards.