Add support for zerocopy `Vec<T>` and IGMP packets (#14)
This PR does two things:

- It adds support for `#[ingot(var_len = "..", zerocopy)]` fields, which
are a `Vec<T>`
- It implements IGMP packets using that new primitive

The proc macro changes are a little awkward! It requires the type `T` to
implement a bunch `ingot_types` traits, as well as `zerocopy` traits.
However, it's hard to do a blanket implementation on all
zerocopy-flavored types, because of trait coherence. Instead, I added a
new `zerocopy_type!` macro which does all of the trait implementations
for you.