Fix blanket impl (#18)
Previously, `zerocopy_type!` didn't work outside of the `ingot_types`
crate, because it implemented `ingot_types::HasView<ObjectSlice<B, $ty>>
for Vec<$ty>`. Outside of the `ingot_types` crate, both the trait and
the type are foreign, so this is not allowed!

This PR adds a blanket `HasView` impl for all `ObjectSlice`, and remove
it from the macro. Doing so requires changing the other blanket impl
back to `Vec<u8>` to avoid collisions, but that was my fault (I made it
overly generic in #14).