Use next-layer type when delegating next_layer (1/) (#24)
Right now, the `next_layer` type when subparsing must always be the same
as our current layer. However, we've talked about changing this in ingot#14.

When we end up supporting heterogenous subparses, it will be more
correct to return the type of the next layer; in the meantime, this
doesn't break anything.