Move hint fallback into `Repeated` / `RepeatedView` (#31)
The existing `NextLayer` implementation requires the `Hint` and `Denom`
to be the same, which prevents heterogeneous `next_layer` parsing. This
is for the sake of the `Repeated` and `RepeatedView` types, where the
next layer may be empty.

This PR moves fallback handling of an empty repeated layer into
`Repeated` and `RepeatedView`, allowing for heterogeneous subparsing
using existing traits.