Remove nested types from `FieldRef` / `FieldMut` (#32)
Previously, there were two owned variants of `FieldRef`. The
`FieldRef::Raw` variant wraps a `&'a Header`, which can be either owned
or borrowed:
```rust
match igmp.source_addrs_ref() {
    ingot_types::FieldRef::Repr(_a) => todo!("owned"),
    ingot_types::FieldRef::Raw(Header::Repr(_a)) => todo!("also owned"),
    ingot_types::FieldRef::Raw(Header::Raw(ips)) => todo!("borrowed"),
}
```

This PR simplifies things by implementing `From<&'a BoxedHeader<T, <T as
HasView<V>>::ViewType>> for FieldRef<'a, T, V>`. This lets us convert
from the `BoxedHeader` to a `FieldRef` directly, instead of needing to
wrap it.

The same is done for `FieldMut`.